Docs improvement
I think this could make it easier to understand. I would also add an example to clarify what the `xss` boolean does, as well as showing the default values from the config.

If you are fine with this I can create an additional PR with those changes.

I think that users who are not comfortable with tests, etc. will not look into those, and you do not want to digg through tests just to figure out what a package does.

Worts scenario, people just don't use the package, which would be sad, as it is a really good one.